move data dumping into utils
