+++ ceteris_paribus/profiles.py
@@ -1,12 +1,8 @@
-import json
 import logging
-import os
 from collections import OrderedDict
 
 import numpy as np
 import pandas as pd
-
-from ceteris_paribus.plots import PLOTS_DIR
 
 
 def individual_variable_profile(explainer, new_observation, y=None, variables=None, grid_points=101,
@@ -178,42 +174,3 @@ class CeterisParibus:
         print('Training data size: {}'.format(self._data.shape[0]))
         print(self.profile)
 
-    def save_profiles(self, profiles, filename):
-        data = self.dump_profiles(profiles)
-        with open(os.path.join(PLOTS_DIR, filename), 'w') as f:
-            f.write("profile = {};".format(json.dumps(data, indent=2, default=self.default)))
-
-    def dump_profiles(self, profiles):
-        data = []
-        for cp_profile in profiles:
-            for i, row in cp_profile.profile.iterrows():
-                data.append(dict(zip(cp_profile.profile.columns, row)))
-        return data
-
-    @staticmethod
-    def default(o):
-        """
-        Workaround for dumping arrays with np.int64 type into json
-        From: https://stackoverflow.com/a/50577730/7828646
-
-        """
-        return int(o)
-
-    def save_observations(self, profiles, filename):
-        data = self.dump_observations(profiles)
-        with open(os.path.join(PLOTS_DIR, filename), 'w') as f:
-            f.write("observation = {};".format(json.dumps(data, indent=2, default=self.default)))
-
-    def dump_observations(self, profiles):
-        data = []
-        for profile in profiles:
-            for i, yhat in enumerate(profile.new_observation_predictions):
-                for var_name in profile.selected_variables:
-                    d = dict(zip(profile._all_variable_names, profile._new_observation.iloc[i]))
-                    d['_vname_'] = var_name
-                    d['_yhat_'] = yhat
-                    d['_label_'] = profile._label
-                    d['_ids_'] = i
-                    d['_y_'] = profile.new_observation_true[i] if profile.new_observation_true is not None else None
-                    data.append(d)
-        return data
